sb_project/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.3
django-bootstrap4==2.3.1
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1